packages/core/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bfg-core"
version = "0.1.0"
description = "Shared library for SCIL Profile and Gr8hub"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.1",
    "python-multipart>=0.0.19",
    "redis[hiredis]>=5.2.0",
    "httpx>=0.28.0",
    "anthropic>=0.42.0",
    "openai>=1.50.0",
    "qdrant-client>=1.12.0",
    "numpy>=2.2.0",
    "structlog>=24.1.0",
    "celery[redis]>=5.4.0",
    "aiosmtplib>=3.0.0",
    "stripe>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28.0",
]

[tool.setuptools.packages.find]
include = ["bfg_core*"]
